Move the messages.go to the proto package. Move all the serialize definitions into messages.go for later use when writing the serialize.go functions. Will not compile until this is finished.
