+++ lib/proto/messages.go
@@ -1,4 +1,4 @@
-package main
+package proto
 
 import (
 	"fmt"
@@ -7,7 +7,6 @@ import (
 	"buildblast/lib/coords"
 	"buildblast/lib/game"
 	"buildblast/lib/mapgen"
-	"buildblast/lib/proto"
 	"buildblast/lib/vmath"
 )
 
@@ -35,7 +34,7 @@ const (
 	MSG_SCOREBOARD_REMOVE                   // CLIENT <--- SERVER
 )
 
-func idToType(id MessageId) Message {
+func IdToType(id MessageId) Message {
 	switch id {
 	case MSG_ENTITY_CREATE:
 		return &MsgEntityCreate{}
@@ -73,7 +72,7 @@ func idToType(id MessageId) Message {
 	panic(fmt.Sprintf("Unknown message recieved from client: %d", id))
 }
 
-func typeToId(m Message) MessageId {
+func TypeToId(m Message) MessageId {
 	switch m.(type) {
 	case *MsgHandshakeReply:
 		return MSG_HANDSHAKE_REPLY
@@ -128,8 +127,8 @@ type MsgHandshakeReply struct {
 func (msg *MsgHandshakeReply) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_HANDSHAKE_REPLY))
-	buf = append(buf, proto.MarshalFloat64(msg.ServerTime)...)
-	buf = append(buf, proto.MarshalString(msg.ClientID)...)
+	buf = append(buf, MarshalFloat64(msg.ServerTime)...)
+	buf = append(buf, MarshalString(msg.ClientID)...)
 	buf = append(buf, msg.PlayerEntityInfo.ToProto()...)
 	return buf
 }
@@ -145,7 +144,7 @@ type MsgHandshakeError struct {
 func (msg *MsgHandshakeError) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_HANDSHAKE_ERROR))
-	buf = append(buf, proto.MarshalString(msg.Message)...)
+	buf = append(buf, MarshalString(msg.Message)...)
 	return buf
 }
 
@@ -156,13 +155,13 @@ func (msg *MsgHandshakeError) FromProto(buf []byte) (int, error) {
 type MsgEntityCreate struct {
 	ID    game.EntityId
 	Kind  game.EntityKind
-	State proto.Proto
+	State Proto
 }
 
 func (msg *MsgEntityCreate) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_ENTITY_CREATE))
-	buf = append(buf, proto.MarshalString(string(msg.ID))...)
+	buf = append(buf, MarshalString(string(msg.ID))...)
 	buf = append(buf, byte(msg.Kind))
 	buf = append(buf, msg.State.ToProto()...)
 	return buf
@@ -175,13 +174,13 @@ func (msg *MsgEntityCreate) FromProto(buf []byte) (int, error) {
 type MsgEntityState struct {
 	ID    game.EntityId
 	Kind  game.EntityKind
-	State proto.Proto
+	State Proto
 }
 
 func (msg *MsgEntityState) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_ENTITY_STATE))
-	buf = append(buf, proto.MarshalString(string(msg.ID))...)
+	buf = append(buf, MarshalString(string(msg.ID))...)
 	buf = append(buf, byte(msg.Kind))
 	buf = append(buf, msg.State.ToProto()...)
 	return buf
@@ -198,7 +197,7 @@ type MsgEntityRemove struct {
 func (msg *MsgEntityRemove) ToProto() []byte {
 	buf := make([]byte, 0, 64)
 	buf = append(buf, byte(MSG_ENTITY_REMOVE))
-	buf = append(buf, proto.MarshalString(string(msg.ID))...)
+	buf = append(buf, MarshalString(string(msg.ID))...)
 	return buf
 }
 
@@ -267,7 +266,7 @@ func (msg *MsgChatSend) ToProto() []byte {
 
 func (msg *MsgChatSend) FromProto(buf []byte) (int, error) {
 	var read int
-	msg.Message, read = proto.UnmarshalString(buf[1:])
+	msg.Message, read = UnmarshalString(buf[1:])
 	return read + 1, nil
 }
 
@@ -279,8 +278,8 @@ type MsgChatBroadcast struct {
 func (msg *MsgChatBroadcast) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_CHAT_BROADCAST))
-	buf = append(buf, proto.MarshalString(msg.DisplayName)...)
-	buf = append(buf, proto.MarshalString(msg.Message)...)
+	buf = append(buf, MarshalString(msg.DisplayName)...)
+	buf = append(buf, MarshalString(msg.Message)...)
 	return buf
 }
 
@@ -322,7 +321,7 @@ type MsgNtpSyncReply struct {
 func (msg *MsgNtpSyncReply) ToProto() []byte {
 	buf := make([]byte, 0, 3*8+1)
 	buf = append(buf, byte(MSG_NTP_SYNC_REPLY))
-	buf = append(buf, proto.MarshalFloat64(msg.ServerTime)...)
+	buf = append(buf, MarshalFloat64(msg.ServerTime)...)
 	return buf
 }
 
@@ -340,10 +339,10 @@ type MsgInventoryState struct {
 func (msg *MsgInventoryState) ToProto() []byte {
 	buf := make([]byte, 0, 1024)
 	buf = append(buf, byte(MSG_INVENTORY_STATE))
-	buf = append(buf, proto.MarshalInt(len(msg.Items))...)
+	buf = append(buf, MarshalInt(len(msg.Items))...)
 	buf = append(buf, msg.Items...)
-	buf = append(buf, proto.MarshalInt(msg.ItemLeft)...)
-	buf = append(buf, proto.MarshalInt(msg.ItemRight)...)
+	buf = append(buf, MarshalInt(msg.ItemLeft)...)
+	buf = append(buf, MarshalInt(msg.ItemRight)...)
 	return buf
 }
 
@@ -365,10 +364,10 @@ func (msg *MsgInventorySelect) FromProto(buf []byte) (int, error) {
 	var read int
 	offset := 1
 	// Client doesn't send the item list
-	value, read = proto.UnmarshalInt(buf[offset:])
+	value, read = UnmarshalInt(buf[offset:])
 	msg.ItemLeft = int(value)
 	offset += read
-	value, read = proto.UnmarshalInt(buf[offset:])
+	value, read = UnmarshalInt(buf[offset:])
 	msg.ItemRight = int(value)
 	offset += read
 	return read, nil
@@ -387,10 +386,10 @@ func (msg *MsgInventoryMove) FromProto(buf []byte) (int, error) {
 	var value int64
 	var read int
 	offset := 1
-	value, read = proto.UnmarshalInt(buf[offset:])
+	value, read = UnmarshalInt(buf[offset:])
 	msg.From = int(value)
 	offset += read
-	value, read = proto.UnmarshalInt(buf[offset:])
+	value, read = UnmarshalInt(buf[offset:])
 	msg.To = int(value)
 	offset += read
 	return read, nil
@@ -404,8 +403,8 @@ type MsgScoreboardAdd struct {
 func (msg *MsgScoreboardAdd) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_SCOREBOARD_ADD))
-	buf = append(buf, proto.MarshalString(msg.Name)...)
-	buf = append(buf, proto.MarshalInt(msg.Score)...)
+	buf = append(buf, MarshalString(msg.Name)...)
+	buf = append(buf, MarshalInt(msg.Score)...)
 	return buf
 }
 
@@ -421,8 +420,8 @@ type MsgScoreboardSet struct {
 func (msg *MsgScoreboardSet) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_SCOREBOARD_SET))
-	buf = append(buf, proto.MarshalString(msg.Name)...)
-	buf = append(buf, proto.MarshalInt(msg.Score)...)
+	buf = append(buf, MarshalString(msg.Name)...)
+	buf = append(buf, MarshalInt(msg.Score)...)
 	return buf
 }
 
@@ -437,7 +436,7 @@ type MsgScoreboardRemove struct {
 func (msg *MsgScoreboardRemove) ToProto() []byte {
 	buf := make([]byte, 0, 256)
 	buf = append(buf, byte(MSG_SCOREBOARD_REMOVE))
-	buf = append(buf, proto.MarshalString(msg.Name)...)
+	buf = append(buf, MarshalString(msg.Name)...)
 	return buf
 }
 
@@ -445,6 +444,175 @@ func (msg *MsgScoreboardRemove) FromProto(buf []byte) (int, error) {
 	panic("FromProto not implemented: MsgScoreboardRemove")
 }
 
-type Message interface {
-	proto.Proto
+type Message interface {}
+
+/*
+func (d Direction) ToProto() []byte {
+	vec3 := d.Vec3()
+	return vec3.ToProto()
 }
+
+func (wc World) ToProto() []byte {
+	vec3 := wc.Vec3()
+	return vec3.ToProto()
+}
+
+func (bc Block) ToProto() []byte {
+	buf := make([]byte, 0, 30)
+	buf = append(buf, proto.MarshalInt(bc.X)...)
+	buf = append(buf, proto.MarshalInt(bc.Y)...)
+	buf = append(buf, proto.MarshalInt(bc.Z)...)
+	return buf
+}
+
+func (bc *Block) FromProto(buf []byte) (int, error) {
+	offset := 0
+	var value int64
+	var read int
+	value, read = proto.UnmarshalInt(buf)
+	bc.X = int(value)
+	offset += read
+	value, read = proto.UnmarshalInt(buf[offset:])
+	bc.Y = int(value)
+	offset += read
+	value, read = proto.UnmarshalInt(buf[offset:])
+	bc.Z = int(value)
+	offset += read
+	return offset, nil
+}
+
+func (cc Chunk) ToProto() []byte {
+	buf := make([]byte, 0, 30)
+	buf = append(buf, proto.MarshalInt(cc.X)...)
+	buf = append(buf, proto.MarshalInt(cc.Y)...)
+	buf = append(buf, proto.MarshalInt(cc.Z)...)
+	return buf
+}
+
+func (oc Offset) ToProto() []byte {
+	buf := make([]byte, 0, 30)
+	buf = append(buf, proto.MarshalInt(oc.X)...)
+	buf = append(buf, proto.MarshalInt(oc.Y)...)
+	buf = append(buf, proto.MarshalInt(oc.Z)...)
+	return buf
+}
+
+func (v *Vec3) ToProto() []byte {
+	buf := make([]byte, 0, 3*8)
+	buf = append(buf, proto.MarshalFloat64(v.X)...)
+	buf = append(buf, proto.MarshalFloat64(v.Y)...)
+	buf = append(buf, proto.MarshalFloat64(v.Z)...)
+	return buf
+}
+
+func (v *Vec3) FromProto(buf []byte) (int, error) {
+	if len(buf) < 3*8 {
+		return 0, errors.New("Buffer too small: Vec3")
+	}
+	v.X, _ = proto.UnmarshalFloat64(buf[0:8])
+	v.Y, _ = proto.UnmarshalFloat64(buf[8:16])
+	v.Z, _ = proto.UnmarshalFloat64(buf[16:24])
+	return 3 * 8, nil
+}
+
+func (es *EntityState) ToProto() []byte {
+	buf := make([]byte, 0, 256)
+	buf = append(buf, proto.MarshalString(string(es.EntityId))...)
+	buf = append(buf, es.Body.ToProto()...)
+	buf = append(buf, proto.MarshalFloat64(es.LastUpdated)...)
+	return buf
+}
+
+func (es *EntityState) FromProto(buf []byte) (int, error) {
+	var idString string
+	var offset int
+	idString, offset = proto.UnmarshalString(buf)
+	es.EntityId = EntityId(idString)
+	read, err := es.Body.FromProto(buf[offset:])
+	offset += read
+	if err != nil {
+		return 0, err
+	}
+	es.LastUpdated, read = proto.UnmarshalFloat64(buf[offset:])
+	offset += read
+	return offset, nil
+}
+
+func (h *Health) ToProto() []byte {
+	return proto.MarshalInt(h.Life)
+}
+
+func (h *Health) FromProto(buf []byte) (int, error) {
+	life, read := proto.UnmarshalInt(buf)
+	h.Life = int(life)
+	return read, nil
+}
+
+func (bs *BioticState) ToProto() []byte {
+	buf := make([]byte, 0, 256)
+	buf = append(buf, bs.EntityState.ToProto()...)
+	buf = append(buf, bs.Health.ToProto()...)
+	return buf
+}
+
+func (bs *BioticState) FromProto(buf []byte) (int, error) {
+	var offset int
+	read, err := bs.EntityState.FromProto(buf)
+	if err != nil {
+		return 0, err
+	}
+	offset, err = bs.Health.FromProto(buf[read:])
+	if err != nil {
+		return 0, err
+	}
+	return read + offset, nil
+}
+
+func (cs *ControlState) ToProto() []byte {
+	panic("ToProto not implemented: ControlState")
+}
+
+func (cs *ControlState) FromProto(buf []byte) (int, error) {
+	value, read := proto.UnmarshalInt(buf)
+	cs.controlFlags = int(value)
+	cs.Lat, _ = proto.UnmarshalFloat64(buf[read+0*8 : read+1*8])
+	cs.Lon, _ = proto.UnmarshalFloat64(buf[read+1*8 : read+2*8])
+	cs.Timestamp, _ = proto.UnmarshalFloat64(buf[read+2*8 : read+3*8])
+	cs.ViewTimestamp, _ = proto.UnmarshalFloat64(buf[read+3*8 : read+4*8])
+	return read + 32, nil
+}
+
+func (b *Body) ToProto() []byte {
+	buf := make([]byte, 0, 5*3*8)
+	buf = append(buf, b.Pos.ToProto()...)
+	buf = append(buf, b.Vel.ToProto()...)
+	buf = append(buf, b.Dir.ToProto()...)
+	buf = append(buf, b.HalfExtents.ToProto()...)
+	buf = append(buf, b.CenterOffset.ToProto()...)
+	return buf
+}
+
+func (b *Body) FromProto(buf []byte) (int, error) {
+	if len(buf) < 5*3*8 {
+		return 0, errors.New("Buffer too small: Body")
+	}
+	size := 3 * 8
+	b.Pos.FromProto(buf[0*size:])
+	b.Vel.FromProto(buf[1*size:])
+	b.Dir.FromProto(buf[2*size:])
+	b.HalfExtents.FromProto(buf[3*size:])
+	b.CenterOffset.FromProto(buf[4*size:])
+	return 5 * 3 * 8, nil
+}
+
+func (wis *WorldItemState) ToProto() []byte {
+	buf := make([]byte, 0, 256)
+	buf = append(buf, wis.EntityState.ToProto()...)
+	buf = append(buf, byte(wis.ItemKind))
+	return buf
+}
+
+func (wis *WorldItemState) FromProto(buf []byte) (int, error) {
+	return 0, nil
+}
+ */
