+++ lib/proto/serialize.go
@@ -1,0 +1,10 @@
+package proto
+
+//import (
+//	"reflect"
+//	"buildblast"
+//)
+
+//func SerializeMessage(msg *main.Message) []byte {
+//
+//}
