+++ lib/vmath/vec.go
@@ -1,10 +1,7 @@
 package vmath
 
 import (
-	"errors"
 	"math"
-
-	"buildblast/lib/proto"
 )
 
 type Vec3 struct {
@@ -74,24 +71,4 @@ func (v *Vec3) DistTo(other *Vec3) float64 {
 
 func (v *Vec3) DistBetween(other *Vec3) float64 {
 	return math.Abs(v.DistTo(other))
-}
-
-// Protocol stuff
-
-func (v *Vec3) ToProto() []byte {
-	buf := make([]byte, 0, 3*8)
-	buf = append(buf, proto.MarshalFloat64(v.X)...)
-	buf = append(buf, proto.MarshalFloat64(v.Y)...)
-	buf = append(buf, proto.MarshalFloat64(v.Z)...)
-	return buf
-}
-
-func (v *Vec3) FromProto(buf []byte) (int, error) {
-	if len(buf) < 3*8 {
-		return 0, errors.New("Buffer too small: Vec3")
-	}
-	v.X, _ = proto.UnmarshalFloat64(buf[0:8])
-	v.Y, _ = proto.UnmarshalFloat64(buf[8:16])
-	v.Z, _ = proto.UnmarshalFloat64(buf[16:24])
-	return 3 * 8, nil
 }
